refactor: moved logic to sub-module
moved mappings logic to a sub-module
